+++ src/aui-form-field/HISTORY.md
@@ -4,5 +4,6 @@
 
 ## @VERSION@
 
+* [AUI-1750](https://issues.liferay.com/browse/AUI-1750) Create time field
 * [AUI-1784](https://issues.liferay.com/browse/AUI-1784) Convert existing form builder fields to separate rendering from edition logic
 * [AUI-1770](https://issues.liferay.com/browse/AUI-1770) Field rendering should be separate from field editing